Re-added the performance test for inline insert, now disabled.
The makefile generation for tests did not require the making of the jni
library and ipc program.  This caused a dependency to be lost, and
builds to not happen in the correct order.  Sometimes we would try to
make the EE before S2 was finished building, which is not possible.  The
change to buildtools.py fixes this.